modify addTransaction with floatingbutton
